created userRepository and moved rent action there
